requirements.txt
beautifulsoup4==4.9.3
Pillow==8.3.2
piexif==1.1.3
python-dateutil==2.8.2
win32-setctime==1.1.0
PyYAML==5.4.1